clean up ui mod structure
