+++ src/ui/mod.rs
@@ -1,19 +1,25 @@
+//! Implementation of the console UI.
+
 use bevy::{input_focus::InputFocus, text::LineHeight};
 
 use crate::prelude::*;
 
+mod buffer;
 mod console;
 mod data;
 mod text;
+mod view;
 
 pub mod prelude {
+    pub use super::buffer::*;
     pub use super::console::*;
     pub use super::data::*;
-    pub use super::text::*;
+    pub use super::text::prelude::*;
+    pub use super::view::*;
 }
 
 pub fn plugin(app: &mut App) {
-    app.add_plugins(console::plugin);
+    app.add_plugins((console::plugin, text::plugin));
     app.init_resource::<InputFocus>();
     app.init_resource::<ConsoleTextPipeline>();
     app.add_message::<ConsoleScrollMsg>();
